requirements.txt
# requirements.txt
sendgrid==6.11.0
flask==3.0.0
pandas==2.1.4
email-validator==2.1.0
python-dotenv==1.0.0